Move ModelDetails page up one level
