+++ src/pages/ModelDetails/generators.js
